[REFACTOR] separate generation and reasoning service
